# docs/openapi.yaml
openapi: 3.0.1
info:
  description: "REST API for managing gym class sessions, memberships, bookings and\
    \ waitlists."
  title: Fitness Class Booking System
  version: "1.0"
servers:
- description: Default server
  url: /
...
paths:
...
  /api/v1/sessions:
    post:
      description: |
        Schedule a new class session with an instructor and room.

        Time intervals use half-open semantics [start, end):
        - 10:00-11:00 and 11:00-12:00 do NOT conflict (adjacent sessions are allowed)
        - 10:00-11:00 and 10:59-11:59 DO conflict (overlapping sessions are rejected)

        If durationMinutes or capacity are omitted, defaults from the class type are used.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassSessionRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: Session created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Malformed or unreadable request body
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: "Conflict (e.g. instructor/room conflict, capacity exceeds\
            \ room, inactive resource)"
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: "Input validation failed (e.g. past time, non-zero seconds)"
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Schedule a new class session
      tags:
      - Sessions
  /api/v1/sessions/browse:
    get:
      description: |
        Query and filter upcoming class sessions with pagination.

        Default behavior (no parameters): returns SCHEDULED sessions for the next 7 days.

        Time range semantics:
        - If neither from nor to is provided, defaults to [now, now + 7 days)
        - from and to must both be provided or both be absent
        - All boundaries use half-open semantics: [from, to), where from is inclusive and to is exclusive
        - Date shorthand (date=YYYY-MM-DD) returns sessions for that entire studio-local day and cannot be combined with from/to
        - Query span cannot exceed 92 days

        Filter behavior:
        - Unknown filter IDs (classTypeId, instructorId, roomId) return 200 with empty page, not 404
        - Filters are combined with AND logic (all must match)
        - availableOnly=true excludes full sessions (bookedCount == capacity)
        - status defaults to SCHEDULED if not provided

        Sort:
        - Only startsAt is supported; defaults to startsAt,asc
        - sort syntax: field,asc or field,desc

        Pagination:
        - page: zero-based page number (default 0)
        - size: results per page, between 1 and 100 (default 20)

        All timestamps in response are in UTC regardless of the offset used in from/to.
      operationId: browse
      parameters:
      - description: "ISO-8601 start of time range (inclusive). Defaults to now. When\
          \ both from and to are omitted, defaults to next 7 days."
        example: 2026-09-24T10:00:00Z
        in: query
        name: from
        required: false
        schema:
          type: string
          format: date-time
      - description: "ISO-8601 end of time range (exclusive). Note: to is exclusive;\
          \ a session starting exactly at to is not included."
        example: 2026-10-01T10:00:00Z
        in: query
        name: to
        required: false
        schema:
          type: string
          format: date-time
      - description: Studio-local date (YYYY-MM-DD) as shorthand for the entire day.
          Interpreted in studio.timezone config. Cannot be combined with from.
        example: 2026-09-23
        in: query
        name: date
        required: false
        schema:
          type: string
      - description: Filter by class type ID. Unknown IDs match nothing and return
          200 with empty page.
        example: 550e8400-e29b-41d4-a716-446655440000
        in: query
        name: classTypeId
        required: false
        schema:
          type: string
          format: uuid
      - description: Filter by instructor ID. Unknown IDs match nothing and return
          200 with empty page.
        example: 550e8400-e29b-41d4-a716-446655440001
        in: query
        name: instructorId
        required: false
        schema:
          type: string
          format: uuid
      - description: Filter by room ID. Unknown IDs match nothing and return 200 with
          empty page.
        example: 550e8400-e29b-41d4-a716-446655440002
        in: query
        name: roomId
        required: false
        schema:
          type: string
          format: uuid
      - description: "If true, exclude sessions where bookedCount == capacity. Default\
          \ is false."
        example: true
        in: query
        name: availableOnly
        required: false
        schema:
          type: boolean
      - description: Filter by session status. Default is SCHEDULED.
        example: SCHEDULED
        in: query
        name: status
        required: false
        schema:
          type: string
      - description: Zero-based page number. Default is 0.
        example: 0
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
      - description: "Number of results per page. Default is 20, max 100."
        example: 20
        in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
      - description: "Sort expression: field,direction (e.g. startsAt,asc). Only startsAt\
          \ is supported. Default is startsAt,asc."
        example: "startsAt,asc"
        in: query
        name: sort
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseClassSessionScheduleResponse"
          description: Sessions retrieved successfully
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseClassSessionScheduleResponse"
          description: "Validation failed (e.g. invalid range, out of range span,\
            \ invalid sort)"
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Browse upcoming fitness class sessions
      tags:
      - Sessions
  /api/v1/sessions/by-recurrence/{recurrenceId}:
    get:
      operationId: getByRecurrenceId
      parameters:
      - in: path
        name: recurrenceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: Sessions retrieved successfully
        "404":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: No sessions found for this recurrence ID
        "422":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: Malformed UUID
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Get all sessions in a recurrence
      tags:
      - Sessions
  /api/v1/sessions/recurring:
    post:
      description: |
        Generate multiple sessions for a recurring weekly pattern (e.g. Mon/Wed/Fri).
        Occurrences are filtered to exclude the past. DST transitions are handled per local timezone.
        Conflicts are detected and either failed (onConflict=FAIL) or skipped (onConflict=SKIP).
      operationId: createRecurring
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringSessionsRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateRecurringSessionsResponse"
          description: Sessions created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Malformed or unreadable request body
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateRecurringSessionsResponse"
          description: Conflict detected with onConflict=FAIL
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateRecurringSessionsResponse"
          description: Input validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Create recurring weekly sessions
      tags:
      - Sessions
  /api/v1/sessions/{id}:
    delete:
      description: |
        Cancel a SCHEDULED session. Cascades to:
        - All BOOKED bookings become CANCELLED with cancellationType: SESSION_CANCELLED
        - All credits are refunded to members (no late-cancel window applies for session cancellation)
        - All WAITING waitlist entries become EXPIRED
        - One notification per affected member

        Bookings already CHECKED_IN or NO_SHOW are not affected.
        Sessions that have already started but not completed can be cancelled.
      operationId: cancel
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSessionRequest"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelSessionResponse"
          description: Session cancelled successfully
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelSessionResponse"
          description: Session not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelSessionResponse"
          description: "Conflict (e.g. session already cancelled, already completed)"
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelSessionResponse"
          description: Input validation failed (e.g. reason exceeds 255 characters)
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Cancel a scheduled class session
      tags:
      - Sessions
    get:
      operationId: getById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: Session retrieved successfully
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: Session not found
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassSessionScheduleResponse"
          description: Malformed UUID
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Get a session by ID
      tags:
      - Sessions
    patch:
      description: |
        Patch an existing SCHEDULED session to change its time, duration, instructor, room, or capacity.

        Only SCHEDULED sessions can be edited. Sessions that have already started, or are COMPLETED/CANCELLED cannot be modified.

        Capacity changes are validated: new capacity cannot drop below the current booked count,
        and cannot exceed the room's maximum capacity.

        When capacity is increased, waiting list members are promoted automatically (atomically within the transaction).

        When time or resources (instructor/room) change, SESSION_RESCHEDULED notifications are written for all members with active bookings.

        classTypeId is immutable and will be rejected if supplied.
      operationId: patch
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchClassSessionRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PatchClassSessionResponse"
          description: Session updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Malformed or unreadable request body
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PatchClassSessionResponse"
          description: "Conflict or state violation (e.g. capacity, time, status,\
            \ concurrency)"
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PatchClassSessionResponse"
          description: Input validation failed or unknown field supplied
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Edit a scheduled class session
      tags:
      - Sessions
components:
  schemas:
...
    CancelSessionRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 255
          minLength: 0
    CancelSessionResponse:
      type: object
      properties:
        availableSpots:
          type: integer
          format: int32
        bookedCount:
          type: integer
          format: int32
        cancelReason:
          type: string
        cancelledAt:
          type: string
          format: date-time
        capacity:
          type: integer
          format: int32
        classTypeId:
          type: string
          format: uuid
        creditsRefunded:
          type: integer
          format: int32
        endsAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        instructorId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        status:
          type: string
    ClassSessionScheduleResponse:
      type: object
      properties:
        availableSpots:
          type: integer
          format: int32
        bookedCount:
          type: integer
          format: int32
        capacity:
          type: integer
          format: int32
        classTypeId:
          type: string
          format: uuid
        endsAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        instructorId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        status:
          type: string
        waitlistCount:
          type: integer
          format: int32
...
    CreateClassSessionRequest:
      type: object
      properties:
        capacity:
          type: integer
          format: int32
        classTypeId:
          type: string
          format: uuid
        durationMinutes:
          type: integer
          format: int32
        instructorId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
      required:
      - classTypeId
      - instructorId
      - roomId
      - startsAt
...
    CreateRecurringSessionsRequest:
      type: object
      properties:
        capacity:
          type: integer
          format: int32
        classTypeId:
          type: string
          format: uuid
        daysOfWeek:
          type: array
          items:
            type: string
        durationMinutes:
          type: integer
          format: int32
        fromDate:
          type: string
          format: date
        instructorId:
          type: string
          format: uuid
        onConflict:
          type: string
        roomId:
          type: string
          format: uuid
        startTimeLocal:
          $ref: "#/components/schemas/LocalTime"
        toDate:
          type: string
          format: date
      required:
      - classTypeId
      - daysOfWeek
      - fromDate
      - instructorId
      - onConflict
      - roomId
      - startTimeLocal
      - toDate
    CreateRecurringSessionsResponse:
      type: object
      properties:
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/RecurrenceConflict"
        created:
          type: array
          items:
            $ref: "#/components/schemas/ClassSessionScheduleResponse"
        recurrenceId:
          type: string
          format: uuid
        skipped:
          type: integer
          format: int32
...
    ErrorEnvelope:
      type: object
      properties:
        code:
          type: string
          enum:
          - MALFORMED_REQUEST
          - NOT_PERMITTED
          - MEMBER_SUSPENDED
          - MEMBER_INACTIVE
          - NOT_FOUND
          - MEMBER_NOT_FOUND
          - MEMBERSHIP_PLAN_NOT_FOUND
          - MEMBERSHIP_NOT_FOUND
          - CLASS_TYPE_NOT_FOUND
          - INSTRUCTOR_NOT_FOUND
          - ROOM_NOT_FOUND
          - SESSION_NOT_FOUND
          - BOOKING_NOT_FOUND
          - WAITLIST_ENTRY_NOT_FOUND
          - CONFLICT
          - CONCURRENT_MODIFICATION
          - DUPLICATE_BOOKING
          - SESSION_FULL
          - SESSION_CANCELLED
          - SESSION_ALREADY_CANCELLED
          - SESSION_NOT_BOOKABLE
          - BOOKING_ALREADY_CANCELLED
          - BOOKING_NOT_CANCELLABLE
          - BOOKING_ALREADY_CHECKED_IN
          - CHECK_IN_WINDOW_NOT_OPEN
          - WAITLIST_ALREADY_JOINED
          - WAITLIST_ALREADY_PROCESSED
          - MEMBERSHIP_ALREADY_ACTIVE
          - CREDITS_INSUFFICIENT
          - MEMBERSHIP_NO_CREDITS
          - OVERLAPPING_BOOKING
          - BOOKING_OVERLAPS_EXISTING
          - LATE_CANCEL_NO_REFUND
          - INSTRUCTOR_SCHEDULE_CONFLICT
          - ROOM_SCHEDULE_CONFLICT
          - CLASS_TYPE_NAME_ALREADY_EXISTS
          - CLASS_TYPE_INACTIVE
          - INSTRUCTOR_EMAIL_ALREADY_EXISTS
          - INSTRUCTOR_INACTIVE
          - INSTRUCTOR_HAS_FUTURE_SESSIONS
          - ROOM_NAME_ALREADY_EXISTS
          - ROOM_INACTIVE
          - ROOM_HAS_FUTURE_SESSIONS
          - SESSION_INSTRUCTOR_CONFLICT
          - SESSION_ROOM_CONFLICT
          - SESSION_CAPACITY_EXCEEDS_ROOM
          - IDEMPOTENCY_KEY_CONFLICT
          - JOB_ALREADY_RUNNING
          - MEMBER_EMAIL_ALREADY_EXISTS
          - MEMBER_ALREADY_SUSPENDED
          - MEMBER_NOT_SUSPENDED
          - PLAN_NAME_ALREADY_EXISTS
          - PLAN_ALREADY_INACTIVE
          - PLAN_INACTIVE
          - MEMBERSHIP_ALREADY_QUEUED
          - MEMBERSHIP_START_BEFORE_CURRENT_EXPIRY
          - MEMBERSHIP_NOT_CANCELLABLE
          - SESSION_NOT_EDITABLE
          - SESSION_ALREADY_STARTED
          - SESSION_CAPACITY_BELOW_BOOKED
          - VALIDATION_FAILED
          - OUT_OF_RANGE
          - INVALID_DATE_RANGE
          - INVALID_PAGINATION
          - INVALID_SORT_FIELD
          - INVALID_FORMAT
          - INVALID_ENUM
          - UNKNOWN_FIELD
          - FUTURE_REQUIRED
          - INVALID_RANGE
          - RECURRENCE_EMPTY
          - INTERNAL_ERROR
        detail:
          type: string
        errors:
          type: array
          description: Per-field validation errors. Present ONLY on 422 responses.
          items:
            $ref: "#/components/schemas/FieldError"
        instance:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        title:
          type: string
        traceId:
          type: string
        type:
          type: string
    FieldError:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: object
...
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
...
    PageMetadata:
      type: object
      description: Pagination metadata
      properties:
        number:
          type: integer
          format: int32
          description: Zero-based page number
          example: 0
        size:
          type: integer
          format: int32
          description: Number of items per page
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 42
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
          example: 3
...
    PageResponseClassSessionScheduleResponse:
      type: object
      description: Paginated response envelope
      properties:
        content:
          type: array
          description: Page content
          items:
            $ref: "#/components/schemas/ClassSessionScheduleResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
    PatchClassSessionRequest:
      type: object
      properties:
        capacity:
          type: integer
          format: int32
          maximum: 500
          minimum: 1
        durationMinutes:
          type: integer
          format: int32
          maximum: 480
          minimum: 1
        instructorId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    PatchClassSessionResponse:
      type: object
      properties:
        availableSpots:
          type: integer
          format: int32
        bookedCount:
          type: integer
          format: int32
        capacity:
          type: integer
          format: int32
        classTypeId:
          type: string
          format: uuid
        endsAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        instructorId:
          type: string
          format: uuid
        promotedCount:
          type: integer
          format: int32
        roomId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        status:
          type: string
        waitlistCount:
          type: integer
          format: int32
...
    RecurrenceConflict:
      type: object
      properties:
        conflictingSessionId:
          type: string
          format: uuid
        occurrenceIndex:
          type: integer
          format: int32
        type:
          type: string
...